Add mock static to floating-avatars.mock.ts  file
