+++ projects/ui-framework/src/lib/mock.const.ts
@@ -1100,51 +1100,6 @@ export const mockAvatar = () =>
     'women',
   ])}/${randomNumber(0, 99)}.jpg`;
 
-export const mockStaticAvatarsLocation = (): StaticAvatarLocation[] => {
-  return [
-    {
-      ballSize: 30,
-      x: 0.12,
-      y: 0.4,
-    },
-    {
-      ballSize: 17,
-      x: 0.15,
-      y: 0.65,
-    },
-    {
-      ballSize: 18,
-      x: 0.3,
-      y: 0.35,
-    },
-    {
-      ballSize: 50,
-      x: 0.3,
-      y: 0.65,
-    },
-    {
-      ballSize: 19,
-      x: 0.7,
-      y: 0.27,
-    },
-    {
-      ballSize: 35,
-      x: 0.65,
-      y: 0.8,
-    },
-    {
-      ballSize: 55,
-      x: 0.77,
-      y: 0.53,
-    },
-    {
-      ballSize: 35,
-      x: 0.92,
-      y: 0.35,
-    },
-  ];
-};
-
 export const mockImage = (width, height) =>
   `https://picsum.photos/id/${randomNumber(0, 99)}/${width}/${height}`;
